refactor: restructure domain and service packages for clarity, scalability, and IDE-friendly icons
