+++ src/main/java/com/henry/universitycourseschedular/controllers/AuthenticationController.java
@@ -4,7 +4,7 @@ import com.henry.universitycourseschedular.constants.StatusCodes;
 import com.henry.universitycourseschedular.models._dto.*;
 import com.henry.universitycourseschedular.models.user.AuthToken;
 import com.henry.universitycourseschedular.repositories.AuthTokenRepository;
-import com.henry.universitycourseschedular.services.AuthenticationService;
+import com.henry.universitycourseschedular.services.auth.AuthenticationService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
