+++ src/main/java/com/henry/universitycourseschedular/controllers/InvitationController.java
@@ -4,7 +4,7 @@ import com.henry.universitycourseschedular.models._dto.DefaultApiResponse;
 import com.henry.universitycourseschedular.models._dto.InviteHodDto;
 import com.henry.universitycourseschedular.models._dto.SuccessfulInviteDto;
 import com.henry.universitycourseschedular.models.invitation.Invitation;
-import com.henry.universitycourseschedular.services.InvitationService;
+import com.henry.universitycourseschedular.services.messaging.InvitationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
