+++ src/main/java/com/henry/universitycourseschedular/controllers/LectureController.java
@@ -3,7 +3,7 @@ package com.henry.universitycourseschedular.controllers;
 import com.henry.universitycourseschedular.models._dto.DefaultApiResponse;
 import com.henry.universitycourseschedular.models._dto.LecturerDto;
 import com.henry.universitycourseschedular.models.core.Lecturer;
-import com.henry.universitycourseschedular.services.LectureService;
+import com.henry.universitycourseschedular.services.core.LectureService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
