+++ src/main/java/com/henry/universitycourseschedular/services/core/DepartmentServiceImpl.java
@@ -1,0 +1,10 @@
+package com.henry.universitycourseschedular.services.core;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Service @RequiredArgsConstructor @Slf4j
+public class DepartmentServiceImpl implements DepartmentService{
+
+}
