+++ src/main/java/com/henry/universitycourseschedular/services/jobs/GACConstraintSolverServiceImpl.java
@@ -1,14 +1,13 @@
-package com.henry.universitycourseschedular.services.impl;
+package com.henry.universitycourseschedular.services.jobs;
 
 import com.henry.universitycourseschedular.models.schedule.ScheduleEntry;
-import com.henry.universitycourseschedular.services.GACConstraintSolverService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service @RequiredArgsConstructor
-public class IGACConstraintSolverService implements GACConstraintSolverService {
+public class GACConstraintSolverServiceImpl implements GACConstraintSolverService {
 
     @Override
     public List<ScheduleEntry> enforceConstraints(List<ScheduleEntry> tentativeSchedule) {
