+++ src/main/java/com/henry/universitycourseschedular/services/jobs/SimulatedAnnealingService.java
@@ -1,4 +1,4 @@
-package com.henry.universitycourseschedular.services;
+package com.henry.universitycourseschedular.services.jobs;
 
 import com.henry.universitycourseschedular.models.schedule.ScheduleEntry;
