+++ src/main/java/com/henry/universitycourseschedular/services/jobs/SimulatedAnnealingServiceImpl.java
@@ -1,7 +1,6 @@
-package com.henry.universitycourseschedular.services.impl;
+package com.henry.universitycourseschedular.services.jobs;
 
 import com.henry.universitycourseschedular.models.schedule.ScheduleEntry;
-import com.henry.universitycourseschedular.services.SimulatedAnnealingService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -9,7 +8,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service @RequiredArgsConstructor @Slf4j
-public class ISimulatedAnnealingService implements SimulatedAnnealingService {
+public class SimulatedAnnealingServiceImpl implements SimulatedAnnealingService {
 
     @Override
     public List<ScheduleEntry> optimize(List<ScheduleEntry> initialSchedule) {
