+++ src/main/java/com/henry/universitycourseschedular/services/jobs/TimetableGeneratorServiceImpl.java
@@ -1,4 +1,4 @@
-package com.henry.universitycourseschedular.services.impl;
+package com.henry.universitycourseschedular.services.jobs;
 
 import com.henry.universitycourseschedular.models._dto.TimetableDTO;
 import com.henry.universitycourseschedular.models.core.Venue;
@@ -8,10 +8,6 @@ import com.henry.universitycourseschedular.models.schedule.TimeSlot;
 import com.henry.universitycourseschedular.repositories.CourseAssignmentRepository;
 import com.henry.universitycourseschedular.repositories.TimeSlotRepository;
 import com.henry.universitycourseschedular.repositories.VenueRepository;
-import com.henry.universitycourseschedular.services.GACConstraintSolverService;
-import com.henry.universitycourseschedular.services.GreedySchedulerService;
-import com.henry.universitycourseschedular.services.SimulatedAnnealingService;
-import com.henry.universitycourseschedular.services.TimetableGeneratorService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -19,7 +15,7 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class ITimetableGeneratorService implements TimetableGeneratorService {
+public class TimetableGeneratorServiceImpl implements TimetableGeneratorService {
     private final GreedySchedulerService greedyScheduler;
     private final SimulatedAnnealingService annealer;
     private final GACConstraintSolverService gac;
